moving away from usr-bin
I'm trying to use WinGet more for installing applications. In reality
it's just a more automated way of stashing binaries in a directory and
updating the path but it is more automated and should help me have a
consistent experience between non-elevated and elevated shells.

- Dropped the '.' from the wezterm config file since it's being
explicitly referenced.
- Added two new setup scripts - one for system-wide tools that has to be
run as admin and one for per-user tools that will have to be run for
non-elevated and elevated users.
- Removed the WezTerm shortcut because I found a way to update the Start
Menu shortcut via Powershell.
- Added Yazi and dependencies. I've been enjoying using it in Linux and
decided to start using it in Windows as well.